+++ Tests/Functional/Controller/ConnectControllerTest.php
@@ -26,11 +26,19 @@ use Symfony\Component\BrowserKit\Client;
 use Symfony\Component\BrowserKit\Cookie;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
+/**
+ * uses FOSUserBundle which itself contains lots of deprecations.
+ *
+ * @group legacy
+ */
 final class ConnectControllerTest extends WebTestCase
 {
     protected function setUp(): void
     {
         static::$class = AppKernel::class;
+        if (!class_exists(\FOS\UserBundle\Model\User::class)) {
+            $this->markTestSkipped('FOSUserBundle not installed.');
+        }
     }
 
     public static function getKernelClass(): string